Making functions static. Preparing to merge to master branch
