+++ src/compare.c
@@ -6,26 +6,10 @@
 #include "predicate.h"
 #include "compare.h"
 
-#ifndef QUANTIFIER_INCREMENT
-#define QUANTIFIER_INCREMENT 4
-#endif
-
-#ifndef QUANTIFIER_DEFAULT
-#define QUANTIFIER_DEFAULT 4
-#endif
-
-#ifndef QUANTIFIER_STACK_INCREMENT
-#define QUANTIFIER_STACK_INCREMENT 4
-#endif
-
-#ifndef QUANTIFIER_STACK_DEFAULT
-#define QUANTIFIER_STACK_DEFAULT 4
-#endif
-
-enum map_status{
+enum map_type{
 	//The variable cannot be mapped
 	RESERVED,
-	//The variable has not yet been mapped
+	//The variable has not been mapped
 	UNMAPPED,
 	//The variable is mapped to a bound variable
 	BOUND_VAR,
@@ -33,10 +17,11 @@ enum map_status{
 	POINTER
 };
 
+//The data stores the status of a quantified variable during sentence comparison
 struct map_entry{
-	enum map_status status;
+	enum map_type type;
 	union{
-		struct{
+		struct {
 			int range_lower;
 			int range_upper;
 		};
@@ -45,426 +30,557 @@ struct map_entry{
 	};
 };
 
-static struct map_entry **s0_map;
-static struct map_entry **s1_map;
-static unsigned int max_quantifiers = QUANTIFIER_DEFAULT;
-static unsigned int stack_size = 1;
-static unsigned int max_stack = QUANTIFIER_STACK_DEFAULT;
+enum seek_status{
+	FAILURE = 0,
+	SUCCESS = 1,
+	CONTINUE = 2
+};
 
-static int sentence_stronger_recursive(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map);
-static int sentence_equivalent_recursive(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map);
+//A linked list which stores a stack of searched subsentences
+//The linked list is stored on the stack!
+struct subsentence_stack{
+	sentence *s0;
+	sentence *s1;
+	struct map_entry *source_map;
+	struct map_entry *dest_map;
+	struct subsentence_stack *parent;
+};
 
-static int sentence_stronger_exact(sentence *s0, sentence *s1);
-static int sentence_equivalent_exact(sentence *s0, sentence *s1);
-static int sentence_trivially_true_exact(sentence *s0);
-static int sentence_trivially_false_exact(sentence *s1);
+static int sentence_stronger_recursive(struct subsentence_stack *parent);
+static int seek_next(struct subsentence_stack *parent);
 
-extern int debug_flag;
-
-void init_quantifier_map(){
+//Later will have better allocation routines
+static struct map_entry *get_map_list(int size){
 	int i;
+	struct map_entry *output;
 
-	s0_map = malloc(sizeof(struct map_entry *)*max_stack);
-	s1_map = malloc(sizeof(struct map_entry *)*max_stack);
-	for(i = 0; i < max_stack; i++){
-		s0_map[i] = malloc(sizeof(struct map_entry)*max_quantifiers);
-		s1_map[i] = malloc(sizeof(struct map_entry)*max_quantifiers);
-	}
-}
-
-void deinit_quantifier_map(){
-	int i;
-
-	for(i = 0; i < max_stack; i++){
-		free(s0_map[i]);
-		free(s1_map[i]);
-	}
-	free(s0_map);
-	free(s1_map);
-}
-
-static void resize_quantifiers(unsigned int new_max_quantifiers){
-	int i;
-
-	for(i = 0; i < max_stack; i++){
-		s0_map[i] = realloc(s0_map[i], sizeof(struct map_entry)*new_max_quantifiers);
-		s1_map[i] = realloc(s1_map[i], sizeof(struct map_entry)*new_max_quantifiers);
-	}
-
-	max_quantifiers = new_max_quantifiers;
-}
-
-static void trivialize_quantifiers(struct map_entry ***map, int num_vars){
-	int i;
-
-	if(num_vars > max_quantifiers){
-		resize_quantifiers(num_vars + QUANTIFIER_INCREMENT - num_vars%QUANTIFIER_INCREMENT);
-	}
-
-	for(i = 0; i < num_vars; i++){
-		(*map)[stack_size - 1][i].status = BOUND_VAR;
-		(*map)[stack_size - 1][i].var_id = i;
-	}
-}
-
-static void resize_stack(unsigned int new_max_stack){
-	int i;
-
-	s0_map = realloc(s0_map, sizeof(struct map_entry *)*new_max_stack);
-	s1_map = realloc(s1_map, sizeof(struct map_entry *)*new_max_stack);
-	for(i = max_stack; i < new_max_stack; i++){
-		s0_map[i] = malloc(sizeof(struct map_entry)*max_quantifiers);
-		s1_map[i] = malloc(sizeof(struct map_entry)*max_quantifiers);
-	}
-
-	max_stack = new_max_stack;
-}
-
-static void add_quantifiers(struct map_entry ***source_map, struct map_entry ***dest_map, int source_lower, int source_upper, int dest_lower, int dest_upper){
-	int i;
-	int max;
-
-	if(source_upper > dest_upper){
-		max = source_upper;
-	} else {
-		max = dest_upper;
-	}
-
-	if(max >= max_quantifiers){
-		resize_quantifiers(max + QUANTIFIER_INCREMENT - max%QUANTIFIER_INCREMENT);
-	}
-
-	for(i = source_lower; i <= source_upper; i++){
-		(*source_map)[stack_size - 1][i].status = UNMAPPED;
-		(*source_map)[stack_size - 1][i].range_lower = dest_lower;
-		(*source_map)[stack_size - 1][i].range_upper = dest_upper;
-	}
-
-	for(i = dest_lower; i <= dest_upper; i++){
-		(*dest_map)[stack_size - 1][i].status = RESERVED;
-	}
-}
-
-static void save_quantifiers(){
-	stack_size++;
-
-	if(stack_size >= max_stack){
-		resize_stack(max_stack + QUANTIFIER_STACK_INCREMENT);
-	}
-
-	memcpy(s0_map[stack_size - 1], s0_map[stack_size - 2], sizeof(struct map_entry)*max_quantifiers);
-	memcpy(s1_map[stack_size - 1], s1_map[stack_size - 2], sizeof(struct map_entry)*max_quantifiers);
-}
-
-static void restore_quantifiers(){
-	stack_size--;
-}
-
-//Copies the current quantifiers into the previous stack frame
-static void copy_quantifiers(){
-	memcpy(s0_map[stack_size - 2], s0_map[stack_size - 1], sizeof(struct map_entry)*max_quantifiers);
-	memcpy(s1_map[stack_size - 2], s1_map[stack_size - 1], sizeof(struct map_entry)*max_quantifiers);
-}
-
-//Checks whether entry0 can be mapped to entry1
-static int check_arguments(struct map_entry entry0, struct map_entry entry1, struct map_entry ***source_map){
-	struct map_entry *source_entry;
-
-	if(entry0.status == BOUND_VAR){
-		source_entry = (*source_map)[stack_size - 1] + entry0.var_id;
-	}
-
-	if(entry0.status == BOUND_VAR && entry1.status == POINTER){
-		if(source_entry->status == UNMAPPED){
-			source_entry->status = POINTER;
-			source_entry->var = entry1.var;
-			return 1;
-		} else if(source_entry->status == POINTER && source_entry->var == entry1.var){
-			return 1;
-		} else {
-			return 0;
-		}
-	} else if(entry0.status == BOUND_VAR && entry1.status == BOUND_VAR){
-		if(source_entry->status == UNMAPPED && entry1.var_id >= source_entry->range_lower && entry1.var_id <= source_entry->range_upper){
-			source_entry->status = BOUND_VAR;
-			source_entry->var_id = entry1.var_id;
-			return 1;
-		} else if(source_entry->status == BOUND_VAR && source_entry->var_id == entry1.var_id){
-			return 1;
-		} else {
-			return 0;
-		}
-	} else if(entry0.status == POINTER && entry1.status == POINTER && entry0.var == entry1.var){
-		return 1;
-	} else {
-		return 0;
-	}
-}
-
-static int stronger_premise_or(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	return sentence_stronger_recursive(s0->child0, s1, source_map, dest_map) && sentence_stronger_recursive(s0->child1, s1, source_map, dest_map);
-}
-
-static int stronger_conclusion_and(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	return sentence_stronger_recursive(s0, s1->child0, source_map, dest_map) && sentence_stronger_recursive(s0, s1->child1, source_map, dest_map);
-}
-
-static int stronger_premise_and(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	save_quantifiers();
-	if(sentence_stronger_recursive(s0->child0, s1, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
-
-	save_quantifiers();
-	if(sentence_stronger_recursive(s0->child1, s1, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
-
-	return 0;
-}
-
-static int stronger_conclusion_or(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	save_quantifiers();
-	if(sentence_stronger_recursive(s0, s1->child0, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
-
-	save_quantifiers();
-	if(sentence_stronger_recursive(s0, s1->child1, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
-
-	return 0;
-}
-
-static int stronger_implies(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	return sentence_stronger_recursive(s1->child0, s0->child0, dest_map, source_map) && sentence_stronger_recursive(s0->child1, s1->child1, source_map, dest_map);
-}
-
-static int stronger_forall(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	int quantifiers_before;
-	int quantifiers_after;
-	int output;
-	int i;
-	sentence *child1;
-
-	quantifiers_before = s1->num_bound_vars;
-	if(s1->type == FORALL){
-		quantifiers_after = s1->child0->num_bound_vars;
-		child1 = s1->child0;
-	} else {
-		quantifiers_after = s1->num_bound_vars;
-		child1 = s1;
-	}
-
-	add_quantifiers(source_map, dest_map, s0->num_bound_vars, s0->child0->num_bound_vars - 1, quantifiers_before, quantifiers_after - 1);
-
-	output = sentence_stronger_recursive(s0->child0, child1, source_map, dest_map);
-
-	for(i = s0->num_bound_vars; i < s0->child0->num_bound_vars; i++){
-		//The quantifiers must have been bound
-		if((*source_map)[stack_size - 1][i].status == UNMAPPED){
-			output = 0;
-			break;
-		}
+	output = malloc(sizeof(struct map_entry)*size);
+	for(i = 0; i < size; i++){
+		output[i].type = RESERVED;
 	}
 
 	return output;
 }
 
-static int stronger_exists(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	int quantifiers_before;
-	int quantifiers_after;
-	int output;
+static void free_map_list(struct map_entry *map){
+	free(map);
+}
+
+static int stronger_premise_or(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0->child0, .s1 = parent->s1, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int stronger_conclusion_and(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0, .s1 = parent->s1->child0, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int stronger_contrapositive(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s1->child0, .s1 = parent->s0->child0, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int stronger_double_negation(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0, .s1 = parent->s1->child0->child0, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int stronger_implies(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s1->child0, .s1 = parent->s0->child0, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int stronger_forall(struct subsentence_stack *parent){
 	int i;
-	sentence *child0;
+	int range_lower;
+	int range_upper;
+	int output;
+	sentence *s1_child;
+	struct subsentence_stack entry;
+	struct map_entry *new_source_map;
+	struct map_entry *new_dest_map;
 
-	quantifiers_before = s0->num_bound_vars;
-	if(s0->type == EXISTS){
-		quantifiers_after = s0->child0->num_bound_vars;
-		child0 = s0->child0;
+	new_source_map = get_map_list(parent->s0->child0->num_bound_vars);
+	memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+
+	if(parent->s1->type == FORALL){
+		s1_child = parent->s1->child0;
 	} else {
-		quantifiers_after = s0->num_bound_vars;
-		child0 = s0;
+		s1_child = parent->s1;
 	}
 
-	add_quantifiers(dest_map, source_map, s1->num_bound_vars, s1->child0->num_bound_vars - 1, quantifiers_before, quantifiers_after - 1);
+	range_lower = parent->s1->num_bound_vars;
+	range_upper = s1_child->num_bound_vars - 1;
 
-	output = sentence_stronger_recursive(child0, s1->child0, source_map, dest_map);
-
-	for(i = s1->num_bound_vars; i < s1->child0->num_bound_vars; i++){
-		if((*dest_map)[stack_size - 1][i].status == UNMAPPED){
-			output = 0;
-			break;
-		}
+	new_dest_map = get_map_list(s1_child->num_bound_vars);
+	memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
+	for(i = range_lower; i <= range_upper; i++){
+		new_dest_map[i].type = RESERVED;
 	}
+
+	for(i = parent->s0->num_bound_vars; i < parent->s0->child0->num_bound_vars; i++){
+		new_source_map[i].type = UNMAPPED;
+		new_source_map[i].range_lower = range_lower;
+		new_source_map[i].range_upper = range_upper;
+	}
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0->child0, .s1 = s1_child, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+	output = sentence_stronger_recursive(&entry);
+	free_map_list(new_source_map);
+	free_map_list(new_dest_map);
 
 	return output;
 }
 
-static int stronger_bicond_both(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	save_quantifiers();
-	if(sentence_equivalent_recursive(s0->child0, s1->child0, source_map, dest_map) && sentence_equivalent_recursive(s0->child1, s1->child1, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
+static int stronger_exists(struct subsentence_stack *parent){
+	int i;
+	int range_lower;
+	int range_upper;
+	int output;
+	sentence *s0_child;
+	struct subsentence_stack entry;
+	struct map_entry *new_source_map;
+	struct map_entry *new_dest_map;
 
-	save_quantifiers();
-	if(sentence_equivalent_recursive(s0->child0, s1->child1, source_map, dest_map) && sentence_equivalent_recursive(s0->child1, s1->child0, source_map, dest_map)){
-		copy_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
+	new_dest_map = get_map_list(parent->s1->child0->num_bound_vars);
+	memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
 
-	save_quantifiers();
-	trivialize_quantifiers(source_map, s1->num_bound_vars);
-	trivialize_quantifiers(dest_map, s1->num_bound_vars);
-	if(sentence_equivalent_recursive(s1->child0, s1->child1, source_map, dest_map)){
-		restore_quantifiers();
-		return 1;
+	if(parent->s0->type == EXISTS){
+		s0_child = parent->s0->child0;
+	} else {
+		s0_child = parent->s0;
 	}
-	restore_quantifiers();
 
-	return 0;
+	range_lower = parent->s0->num_bound_vars;
+	range_upper = s0_child->num_bound_vars - 1;
+
+	new_source_map = get_map_list(s0_child->num_bound_vars);
+	memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+	for(i = range_lower; i <= range_upper; i++){
+		new_source_map[i].type = RESERVED;
+	}
+
+	for(i = parent->s1->num_bound_vars; i < parent->s1->child0->num_bound_vars; i++){
+		new_dest_map[i].type = UNMAPPED;
+		new_dest_map[i].range_lower = range_lower;
+		new_dest_map[i].range_upper = range_upper;
+	}
+
+	entry = (struct subsentence_stack) {.s0 = s0_child, .s1 = parent->s1->child0, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+	output = sentence_stronger_recursive(&entry);
+	free_map_list(new_dest_map);
+	free_map_list(new_source_map);
+
+	return output;
 }
 
-static int stronger_bicond_conclusion(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	save_quantifiers();
-	trivialize_quantifiers(source_map, s1->num_bound_vars);
-	trivialize_quantifiers(dest_map, s1->num_bound_vars);
-	if(sentence_equivalent_recursive(s1->child0, s1->child1, source_map, dest_map)){
-		restore_quantifiers();
-		return 1;
-	}
-	restore_quantifiers();
+static int stronger_bicond_both(struct subsentence_stack *parent){
+	struct subsentence_stack entry0;
+	struct subsentence_stack entry1;
 
-	return 0;
+	entry0 = (struct subsentence_stack) {.s0 = parent->s0->child0, .s1 = parent->s1->child0, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+	entry1 = (struct subsentence_stack) {.s0 = parent->s0->child0, .s1 = parent->s1->child1, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	//Remember, these aren't the final checks
+	//We seek back up the tree and check the other implications once we reach a leaf node
+	return sentence_stronger_recursive(&entry0) || sentence_stronger_recursive(&entry1);
 }
 
-static int stronger_relation(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
+static int bind_arguments(struct map_entry entry0, struct map_entry entry1, struct map_entry *source_map){
+	struct map_entry source_entry;
+
+	if(entry0.type == BOUND_VAR){
+		source_entry = source_map[entry0.var_id];
+		if(source_entry.type == BOUND_VAR){
+			return entry1.type == BOUND_VAR && entry1.var_id == source_entry.var_id;
+		} else if(source_entry.type == POINTER){
+			return entry1.type == POINTER && entry1.var == source_entry.var;
+		} else if(source_entry.type == UNMAPPED){
+			if(entry1.type == BOUND_VAR && entry1.var_id >= source_entry.range_lower && entry1.var_id <= source_entry.range_upper){
+				source_map[entry0.var_id].type = BOUND_VAR;
+				source_map[entry0.var_id].var_id = entry1.var_id;
+				return 1;
+			} else if(entry1.type == POINTER){
+				source_map[entry0.var_id].type = POINTER;
+				source_map[entry0.var_id].var = entry1.var;
+				return 1;
+			} else {
+				return 0;
+			}
+		} else {
+			return 0;
+		}
+	} else if(entry0.type == POINTER && entry1.type == POINTER){
+		return entry0.var == entry1.var;
+	} else {
+		return 0;
+	}
+}
+
+static int stronger_relation(struct subsentence_stack *parent){
 	struct map_entry s0_entry;
 	struct map_entry s1_entry;
 
-	if(s0->relation_data != s1->relation_data){
+	if(parent->s0->relation_data != parent->s1->relation_data){
 		return 0;
 	}
 
-	if(s0->is_bound0){
-		s0_entry.status = BOUND_VAR;
-		s0_entry.var_id = s0->var0_id;
+	if(parent->s0->is_bound0){
+		s0_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s0->var0_id};
 	} else {
-		s0_entry.status = POINTER;
-		s0_entry.var = s0->var0;
+		s0_entry = (struct map_entry) {.type = POINTER, .var = parent->s0->var0};
 	}
 
-	if(s1->is_bound0){
-		s1_entry.status = BOUND_VAR;
-		s1_entry.var_id = s1->var0_id;
+	if(parent->s1->is_bound0){
+		s1_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s1->var0_id};
 	} else {
-		s1_entry.status = POINTER;
-		s1_entry.var = s1->var0;
+		s1_entry = (struct map_entry) {.type = POINTER, .var = parent->s1->var0};
 	}
 
-	if(!check_arguments(s0_entry, s1_entry, source_map) && !check_arguments(s1_entry, s0_entry, dest_map)){
+	if(!bind_arguments(s0_entry, s1_entry, parent->source_map) && !bind_arguments(s1_entry, s0_entry, parent->dest_map)){
 		return 0;
 	}
 
-	if(s0->is_bound1){
-		s0_entry.status = BOUND_VAR;
-		s0_entry.var_id = s0->var1_id;
+	if(parent->s0->is_bound1){
+		s0_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s0->var1_id};
 	} else {
-		s0_entry.status = POINTER;
-		s0_entry.var = s0->var1;
+		s0_entry = (struct map_entry) {.type = POINTER, .var = parent->s0->var1};
 	}
 
-	if(s1->is_bound1){
-		s1_entry.status = BOUND_VAR;
-		s1_entry.var_id = s1->var1_id;
+	if(parent->s1->is_bound1){
+		s1_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s1->var1_id};
 	} else {
-		s1_entry.status = POINTER;
-		s1_entry.var = s1->var1;
+		s1_entry = (struct map_entry) {.type = POINTER, .var = parent->s1->var1};
 	}
 
-	return check_arguments(s0_entry, s1_entry, source_map) || check_arguments(s1_entry, s0_entry, dest_map);
+	if(!bind_arguments(s0_entry, s1_entry, parent->source_map) && !bind_arguments(s1_entry, s0_entry, parent->dest_map)){
+		return 0;
+	}
+
+	return seek_next(parent);
 }
 
-static int stronger_proposition(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
+static int stronger_proposition(struct subsentence_stack *parent){
 	struct map_entry s0_entry;
 	struct map_entry s1_entry;
 	int i;
 
-	if(s0->is_bound != s1->is_bound){
+	if(parent->s0->is_bound != parent->s1->is_bound){
 		return 0;
 	}
-	if(s0->is_bound && s0->definition_id != s1->definition_id){
+	if(parent->s0->is_bound && parent->s0->definition_id != parent->s1->definition_id){
 		return 0;
 	}
-	if(!s0->is_bound && s0->definition_data != s1->definition_data){
+	if(!parent->s0->is_bound && parent->s0->definition_data != parent->s1->definition_data){
 		return 0;
 	}
-	if(s0->num_args != s1->num_args){
+	if(parent->s0->num_args != parent->s1->num_args){
 		return 0;
 	}
 
-	for(i = 0; i < s0->num_args; i++){
-		if(s0->proposition_args[i].is_bound){
-			s0_entry.status = BOUND_VAR;
-			s0_entry.var_id = s0->proposition_args[i].var_id;
+	for(i = 0; i < parent->s0->num_args; i++){
+		if(parent->s0->proposition_args[i].is_bound){
+			s0_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s0->proposition_args[i].var_id};
 		} else {
-			s0_entry.status = POINTER;
-			s0_entry.var = s0->proposition_args[i].var;
+			s0_entry = (struct map_entry) {.type = POINTER, .var = parent->s0->proposition_args[i].var};
 		}
 
-		if(s1->proposition_args[i].is_bound){
-			s1_entry.status = BOUND_VAR;
-			s1_entry.var_id = s1->proposition_args[i].var_id;
+		if(parent->s1->proposition_args[i].is_bound){
+			s1_entry = (struct map_entry) {.type = BOUND_VAR, .var_id = parent->s1->proposition_args[i].var_id};
 		} else {
-			s1_entry.status = POINTER;;
-			s1_entry.var = s1->proposition_args[i].var;
+			s1_entry = (struct map_entry) {.type = POINTER, .var = parent->s1->proposition_args[i].var};
 		}
 
-		if(!check_arguments(s0_entry, s1_entry, source_map) && !check_arguments(s1_entry, s0_entry, dest_map)){
+		if(!bind_arguments(s0_entry, s1_entry, parent->source_map) && !bind_arguments(s1_entry, s0_entry, parent->dest_map)){
 			return 0;
 		}
 	}
 
+	return seek_next(parent);
+}
+
+static int stronger_and_or(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+	struct map_entry *new_source_map;
+	struct map_entry *new_dest_map;
+
+	new_source_map = get_map_list(parent->s0->num_bound_vars);
+	new_dest_map = get_map_list(parent->s1->num_bound_vars);
+
+	if(parent->s0->type == AND){
+
+		memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+		memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
+
+		entry = (struct subsentence_stack) {.s0 = parent->s0->child0, .s1 = parent->s1, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+
+		if(sentence_stronger_recursive(&entry)){
+			free_map_list(new_source_map);
+			free_map_list(new_dest_map);
+			return 1;
+		}
+
+		memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+		memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
+
+		entry = (struct subsentence_stack) {.s0 = parent->s0->child1, .s1 = parent->s1, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+		if(sentence_stronger_recursive(&entry)){
+			free_map_list(new_source_map);
+			free_map_list(new_dest_map);
+			return 1;
+		}
+	}
+
+	if(parent->s1->type == OR){
+
+		memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+		memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
+
+		entry = (struct subsentence_stack) {.s0 = parent->s0, .s1 = parent->s1->child0, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+		if(sentence_stronger_recursive(&entry)){
+			free_map_list(new_source_map);
+			free_map_list(new_dest_map);
+			return 1;
+		}
+
+		memcpy(new_source_map, parent->source_map, sizeof(struct map_entry)*parent->s0->num_bound_vars);
+		memcpy(new_dest_map, parent->dest_map, sizeof(struct map_entry)*parent->s1->num_bound_vars);
+
+		entry = (struct subsentence_stack) {.s0 = parent->s0, .s1 = parent->s1->child1, .parent = parent, .source_map = new_source_map, .dest_map = new_dest_map};
+		if(sentence_stronger_recursive(&entry)){
+			free_map_list(new_source_map);
+			free_map_list(new_dest_map);
+			return 1;
+		}
+	}
+
+	free_map_list(new_source_map);
+	free_map_list(new_dest_map);
+	return 0;
+}
+
+//Recursively check if parent->s0 is stronger than parent->s1
+static int sentence_stronger_recursive(struct subsentence_stack *parent){
+	if(parent->s0->type == OR){
+		if(stronger_premise_or(parent)) return 1;
+	} else if(parent->s1->type == AND){
+		if(stronger_conclusion_and(parent)) return 1;
+	} else if(sentence_trivially_false(parent->s0)){
+		if(seek_next(parent)) return 1;
+	} else if(sentence_trivially_true(parent->s1)){
+		if(seek_next(parent)) return 1;
+	} else if(parent->s0->type == NOT && parent->s1->type == NOT){
+		if(stronger_contrapositive(parent)) return 1;
+	} else if(parent->s1->type == NOT && parent->s1->child0->type == NOT){
+		if(stronger_double_negation(parent)) return 1;
+	} else if(parent->s0->type == IMPLIES && parent->s1->type == IMPLIES){
+		if(stronger_implies(parent)) return 1;
+	} else if(parent->s0->type == FORALL){
+		if(stronger_forall(parent)) return 1;
+	} else if(parent->s1->type == EXISTS){
+		if(stronger_exists(parent)) return 1;
+	} else if(parent->s0->type == BICOND && parent->s1->type == BICOND){
+		if(stronger_bicond_both(parent)) return 1;
+	} else if(parent->s0->type == RELATION && parent->s1->type == RELATION){
+		if(stronger_relation(parent)) return 1;
+	} else if(parent->s0->type == PROPOSITION && parent->s1->type == PROPOSITION){
+		if(stronger_proposition(parent)) return 1;
+	}
+	if(parent->s0->type == AND || parent->s1->type == OR){
+		return stronger_and_or(parent);
+	} else {
+		return 0;
+	}
+}
+
+static int seek_premise_or(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0->child1, .s1 = parent->s1, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int seek_conclusion_and(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0, .s1 = parent->s1->child1, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static int seek_implies(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+
+	entry = (struct subsentence_stack) {.s0 = parent->s0->child1, .s1 = parent->s1->child1, .parent = parent, .source_map = parent->source_map, .dest_map = parent->dest_map};
+
+	return sentence_stronger_recursive(&entry);
+}
+
+static enum seek_status seek_bicond_both(struct subsentence_stack *parent, struct subsentence_stack *child){
+	struct subsentence_stack entry;
+
+	if(child->s0 == parent->s0->child0 && child->s1 == parent->s1->child0){
+		entry = (struct subsentence_stack) {.s0 = parent->s1->child0, .s1 = parent->s0->child0, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	} else if(child->s0 == parent->s1->child0 && child->s1 == parent->s0->child0){
+		//We must take care to swap the source and destination maps back to
+		//what they were before. This is because they were switched by the
+		//above case which was invoked earlier in the tree
+		entry = (struct subsentence_stack) {.s0 = parent->s0->child1, .s1 = parent->s1->child1, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	} else if(child->s0 == parent->s0->child1 && child->s1 == parent->s1->child1){
+		//Once again, we must swap the source and destination
+		entry = (struct subsentence_stack) {.s0 = parent->s1->child1, .s1 = parent->s0->child1, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	} else if(child->s0 == parent->s0->child0 && child->s1 == parent->s1->child1){
+		entry = (struct subsentence_stack) {.s0 = parent->s1->child1, .s1 = parent->s0->child0, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	} else if(child->s0 == parent->s1->child1 && child->s1 == parent->s0->child0){
+		entry = (struct subsentence_stack) {.s0 = parent->s0->child1, .s1 = parent->s1->child0, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	} else if(child->s0 == parent->s0->child1 && child->s1 == parent->s1->child0){
+		entry = (struct subsentence_stack) {.s0 = parent->s1->child0, .s1 = parent->s0->child1, .parent = parent, .source_map = parent->dest_map, .dest_map = parent->source_map};
+		return sentence_stronger_recursive(&entry);
+	}
+
+	return CONTINUE;
+}
+
+static int seek_next(struct subsentence_stack *parent){
+	struct subsentence_stack entry;
+	struct subsentence_stack *child;
+	struct map_entry *leaf_source_map;
+	struct map_entry *leaf_dest_map;
+	struct map_entry *temp_map;
+	int i;
+	enum seek_status output;
+
+	leaf_source_map = parent->source_map;
+	leaf_dest_map = parent->dest_map;
+
+	while(parent->parent != NULL){
+		child = parent;
+		parent = child->parent;
+
+		//We must return to a node in the tree while keeping track of previously
+		//bound variables
+		entry = *parent;
+		entry.source_map = leaf_source_map;
+		entry.dest_map = leaf_dest_map;
+
+		if(parent->s0->type == OR && parent->s0->child0 == child->s0){
+			return seek_premise_or(&entry);
+		} else if(parent->s1->type == AND && parent->s1->child0 == child->s1){
+			return seek_conclusion_and(&entry);
+		} else if(parent->s0->type == NOT && parent->s1->type == NOT){
+			//Evaluating contrapositives swaps the source and
+			//destination maps, so we must swap them back
+			//before continuing up the tree
+			temp_map = leaf_source_map;
+			leaf_source_map = leaf_dest_map;
+			leaf_dest_map = temp_map;
+		} else if(parent->s0->type == IMPLIES && parent->s1->type == IMPLIES && child->s0 == parent->s1->child0 && child->s1 == parent->s0->child0){
+			//While evaluating the implication, the source and
+			//destination maps get swapped. We swap them back
+			//before continuing down the tree
+			temp_map = entry.source_map;
+			entry.source_map = entry.dest_map;
+			entry.dest_map = temp_map;
+
+			return seek_implies(&entry);
+		//If we seek past a quantifier, we need to check all quantifiers got bound
+		//Otherwise the user could prove the existence of an object
+		} else if(parent->s0->type == FORALL && child->s0 == parent->s0->child0){
+			for(i = parent->s0->num_bound_vars; i < child->s0->num_bound_vars; i++){
+				if(leaf_source_map[i].type == UNMAPPED){
+					return 0;
+				}
+				if(leaf_source_map[i].type == RESERVED){
+					custom_error(1, "Internal Error: The source and destination maps are mismatched\n", 1);
+				}
+			}
+		} else if(parent->s1->type == EXISTS && child->s1 == parent->s1->child0){
+			for(i = parent->s1->num_bound_vars; i < child->s1->num_bound_vars; i++){
+				if(leaf_dest_map[i].type == UNMAPPED){
+					return 0;
+				}
+				if(leaf_dest_map[i].type == RESERVED){
+					custom_error(1, "Internal Error: The source and destination maps are mismatched\n", 1);
+				}
+			}
+		} else if(parent->s0->type == BICOND && parent->s1->type == BICOND){
+			output = seek_bicond_both(&entry, child);
+			if(output != CONTINUE){
+				return output;
+			} else {
+				//While evaluating the biconditional implication, the source and
+				//destination maps get swapped. We swap them back before seeking
+				//farther up the tree.
+				temp_map = leaf_source_map;
+				leaf_source_map = leaf_dest_map;
+				leaf_dest_map = temp_map;
+			}
+		}
+	}
+
+	//We made it all the way through the tree! Yay!
+	//This means that the implication has been verified
 	return 1;
 }
 
-static int sentence_equivalent_recursive(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	return sentence_stronger_recursive(s0, s1, source_map, dest_map) && sentence_stronger_recursive(s1, s0, dest_map, source_map);
+int sentence_stronger(sentence *s0, sentence *s1){
+	struct subsentence_stack root;
+	struct map_entry *source_map;
+	struct map_entry *dest_map;
+	int i;
+	int output;
+
+	if(s0->num_bound_vars != s1->num_bound_vars){
+		return 0;
+	}
+
+	source_map = get_map_list(s0->num_bound_vars);
+	dest_map = get_map_list(s1->num_bound_vars);
+
+	for(i = 0; i < s0->num_bound_vars; i++){
+		source_map[i] = (struct map_entry) {.type = BOUND_VAR, .var_id = i};
+		dest_map[i] = (struct map_entry) {.type = RESERVED};
+	}
+
+	root = (struct subsentence_stack) {.s0 = s0, .s1 = s1, .parent = NULL, .source_map = source_map, .dest_map = dest_map};
+	output = sentence_stronger_recursive(&root);
+
+	free_map_list(source_map);
+	free_map_list(dest_map);
+	return output;
+}
+
+int sentence_equivalent(sentence *s0, sentence *s1){
+	return sentence_stronger(s0, s1) && sentence_stronger(s1, s0);
 }
 
 int sentence_trivially_true(sentence *s){
 	if(s->type == IMPLIES){
-		save_quantifiers();
-		trivialize_quantifiers(&s0_map, s->num_bound_vars);
-		trivialize_quantifiers(&s1_map, s->num_bound_vars);
-		if(sentence_stronger_recursive(s->child0, s->child1, &s0_map, &s1_map)){
-			restore_quantifiers();
-			return 1;
-		}
-		restore_quantifiers();
-		return 0;
+		return sentence_stronger(s->child0, s->child1);
 	} else if(s->type == BICOND){
-		save_quantifiers();
-		trivialize_quantifiers(&s0_map, s->num_bound_vars);
-		trivialize_quantifiers(&s1_map, s->num_bound_vars);
-		if(sentence_equivalent_recursive(s->child0, s->child1, &s0_map, &s1_map)){
-			restore_quantifiers();
-			return 1;
-		}
-		restore_quantifiers();
-		return 0;
+		return sentence_equivalent(s->child0, s->child1);
 	} else if(s->type == AND){
 		return sentence_trivially_true(s->child0) && sentence_trivially_true(s->child1);
 	} else if(s->type == OR){
@@ -490,15 +606,15 @@ int sentence_trivially_false(sentence *s){
 		return sentence_trivially_true(s->child0) && sentence_trivially_false(s->child1);
 	} else if(s->type == BICOND){
 		true0 = sentence_trivially_true(s->child0);
-		true1 = sentence_trivially_true(s->child1);
 		false0 = sentence_trivially_false(s->child0);
+		true1 = sentence_trivially_true(s->child1);
 		false1 = sentence_trivially_false(s->child1);
 		return (true0 && false1) || (false0 && true1);
 	} else if(s->type == AND){
 		return sentence_trivially_false(s->child0) || sentence_trivially_false(s->child1);
 	} else if(s->type == OR){
 		return sentence_trivially_false(s->child0) && sentence_trivially_false(s->child1);
-	} else if(s->type == FORALL || s->type == EXISTS){
+	} else if(s->type == EXISTS){
 		return sentence_trivially_false(s->child0);
 	} else if(s->type == NOT){
 		return sentence_trivially_true(s->child0);
@@ -509,218 +625,3 @@ int sentence_trivially_false(sentence *s){
 	}
 }
 
-static int sentence_stronger_recursive(sentence *s0, sentence *s1, struct map_entry ***source_map, struct map_entry ***dest_map){
-	if(s0->type == OR){
-		return stronger_premise_or(s0, s1, source_map, dest_map);
-	} else if(s1->type == AND){
-		return stronger_conclusion_and(s0, s1, source_map, dest_map);
-	} else if(sentence_trivially_false(s0)){
-		return 1;
-	} else if(sentence_trivially_true(s1)){
-		return 1;
-	} else if(s0->type == NOT && s1->type == NOT){
-		return sentence_stronger_recursive(s1->child0, s0->child0, dest_map, source_map);
-	} else if(s1->type == NOT && s1->child0->type == NOT){
-		return sentence_stronger_recursive(s0, s1->child0->child0, source_map, dest_map);
-	} else if(s0->type == IMPLIES && s1->type == IMPLIES){
-		return stronger_implies(s0, s1, source_map, dest_map);
-	} else if(s0->type == FORALL){
-		return stronger_forall(s0, s1, source_map, dest_map);
-	} else if(s1->type == EXISTS){
-		return stronger_exists(s0, s1, source_map, dest_map);
-	} else if(s0->type == BICOND && s1->type == BICOND){
-		return stronger_bicond_both(s0, s1, source_map, dest_map);
-	} else if(s1->type == BICOND){
-		return stronger_bicond_conclusion(s0, s1, source_map, dest_map);
-	} else if(s0->type == RELATION && s1->type == RELATION){
-		return stronger_relation(s0, s1, source_map, dest_map);
-	} else if(s0->type == PROPOSITION && s1->type == PROPOSITION){
-		return stronger_proposition(s0, s1, source_map, dest_map);
-	} else {
-		if(s0->type == AND && stronger_premise_and(s0, s1, source_map, dest_map)){
-			return 1;
-		}
-		if(s1->type == OR && stronger_conclusion_or(s0, s1, source_map, dest_map)){
-			return 1;
-		}
-		return 0;
-	}
-}
-
-int sentence_stronger(sentence *s0, sentence *s1){
-	if(s0->num_bound_vars != s1->num_bound_vars || s0->num_bound_props != s1->num_bound_props){
-		return 0;
-	}
-
-	trivialize_quantifiers(&s0_map, s0->num_bound_vars);
-	trivialize_quantifiers(&s1_map, s1->num_bound_vars);
-	return sentence_stronger_exact(s0, s1) || sentence_stronger_recursive(s0, s1, &s0_map, &s1_map);
-	//return sentence_stronger_recursive(s0, s1, &s0_map, &s1_map);
-}
-
-int sentence_equivalent(sentence *s0, sentence *s1){
-	return sentence_equivalent_exact(s0, s1) || (sentence_stronger(s0, s1) && sentence_stronger(s1, s0));
-}
-
-static int sentence_stronger_exact(sentence *s0, sentence *s1){
-	int i;
-
-	if(s0->type == OR){
-		return sentence_stronger_exact(s0->child0, s1) && sentence_stronger_exact(s0->child1, s1);
-	} else if(s1->type == AND){
-		return sentence_stronger_exact(s0, s1->child0) && sentence_stronger_exact(s0, s1->child1);
-	} else if(sentence_trivially_false_exact(s0)){
-		return 1;
-	} else if(sentence_trivially_true_exact(s1)){
-		return 1;
-	} else if(s0->type == NOT && s1->type == NOT){
-		return sentence_stronger_exact(s1->child0, s0->child0);
-	} else if(s1->type == NOT && s1->child0->type == NOT){
-		return sentence_stronger_exact(s0, s1->child0->child0);
-	} else if(s0->type == IMPLIES && s1->type == IMPLIES){
-		return sentence_stronger_exact(s1->child0, s0->child0) && sentence_stronger_exact(s0->child1, s1->child1);
-	} else if(s0->type == FORALL && s1->type == FORALL){
-		return sentence_stronger_exact(s0->child0, s1->child0);
-	} else if(s0->type == EXISTS && s1->type == EXISTS){
-		return sentence_stronger_exact(s0->child0, s1->child0);
-	} else if(s0->type == BICOND && s1->type == BICOND){
-		if(sentence_equivalent_exact(s0->child0, s1->child0) && sentence_equivalent_exact(s0->child1, s1->child1)){
-			return 1;
-		}
-		if(sentence_equivalent_exact(s0->child0, s1->child1) && sentence_equivalent_exact(s0->child1, s1->child0)){
-			return 1;
-		}
-		//Used to also check here if s0->child0 and s0->child1 are equivalent
-		//But since it's an implication, we only care about s1
-		if(sentence_equivalent_exact(s1->child0, s1->child1)){
-			return 1;
-		}
-		return 0;
-	} else if(s1->type == BICOND){
-		return sentence_equivalent_exact(s1->child0, s1->child1);
-	} else if(s0->type == RELATION && s1->type == RELATION){
-		if(s0->relation_data != s1->relation_data){
-			return 0;
-		}
-		if(s0->is_bound0 != s1->is_bound0){
-			return 0;
-		}
-		if(s0->is_bound1 != s1->is_bound1){
-			return 0;
-		}
-		if(s0->is_bound0){
-			if(s0->var0_id != s1->var0_id){
-				return 0;
-			}
-		} else {
-			if(s0->var0 != s1->var0){
-				return 0;
-			}
-		}
-		if(s0->is_bound1){
-			if(s0->var1_id != s1->var1_id){
-				return 0;
-			}
-		} else {
-			if(s0->var1 != s1->var1){
-				return 0;
-			}
-		}
-		return 1;
-	} else if(s0->type == PROPOSITION && s1->type == PROPOSITION){
-		if(s0->is_bound != s1->is_bound){
-			return 0;
-		}
-		if(s0->is_bound){
-			if(s0->definition_id != s1->definition_id){
-				return 0;
-			}
-		} else {
-			if(s0->definition_data != s1->definition_data){
-				return 0;
-			}
-		}
-		if(s0->num_args != s1->num_args){
-			return 0;
-		}
-		for(i = 0; i < s0->num_args; i++){
-			if(s0->proposition_args[i].is_bound != s1->proposition_args[i].is_bound){
-				return 0;
-			}
-			if(s0->proposition_args[i].is_bound){
-				if(s0->proposition_args[i].var_id != s1->proposition_args[i].var_id){
-					return 0;
-				}
-			} else {
-				if(s0->proposition_args[i].var != s1->proposition_args[i].var){
-					return 0;
-				}
-			}
-		}
-		return 1;
-	} else {
-		if(s0->type == AND && (sentence_stronger_exact(s0->child0, s1) || sentence_stronger_exact(s0->child1, s1))){
-			return 1;
-		}
-		if(s1->type == OR && (sentence_stronger_exact(s0, s1->child0) || sentence_stronger_exact(s0, s1->child1))){
-			return 1;
-		}
-		return 0;
-	}
-}
-
-//Determine if two sentences imply each other
-static int sentence_equivalent_exact(sentence *s0, sentence *s1){
-	return sentence_stronger_exact(s0, s1) && sentence_stronger_exact(s1, s0);
-}
-
-//Determine if a sentence is trivially true
-static int sentence_trivially_true_exact(sentence *s){
-	if(s->type == IMPLIES){
-		return sentence_stronger_exact(s->child0, s->child1);
-	} else if(s->type == BICOND){
-		return sentence_equivalent_exact(s->child0, s->child1);
-	} else if(s->type == AND){
-		return sentence_trivially_true_exact(s->child0) && sentence_trivially_true_exact(s->child1);
-	} else if(s->type == OR){
-		return sentence_trivially_true_exact(s->child0) || sentence_trivially_true_exact(s->child1);
-	} else if(s->type == FORALL){
-		return sentence_trivially_true_exact(s->child0);
-	} else if(s->type == NOT){
-		return sentence_trivially_false_exact(s->child0);
-	} else if(s->type == TRUE){
-		return 1;
-	} else {
-		return 0;
-	}
-}
-
-//Determine if a sentence is trivially false
-static int sentence_trivially_false_exact(sentence *s){
-	int true0;
-	int true1;
-	int false0;
-	int false1;
-
-	if(s->type == IMPLIES){
-		return sentence_trivially_true_exact(s->child0) && sentence_trivially_false_exact(s->child1);
-	} else if(s->type == BICOND){
-		true0 = sentence_trivially_true_exact(s->child0);
-		false0 = sentence_trivially_false_exact(s->child0);
-		true1 = sentence_trivially_true_exact(s->child1);
-		false1 = sentence_trivially_false_exact(s->child1);
-		return (true0 && false1) || (false0 && true1);
-	} else if(s->type == AND){
-		return sentence_trivially_false_exact(s->child0) || sentence_trivially_false_exact(s->child1);
-	} else if(s->type == OR){
-		return sentence_trivially_false_exact(s->child0) && sentence_trivially_false_exact(s->child1);
-	} else if(s->type == EXISTS || s->type == FORALL){
-		return sentence_trivially_false_exact(s->child0);
-	} else if(s->type == NOT){
-		return sentence_trivially_true_exact(s->child0);
-	} else if(s->type == FALSE){
-		return 1;
-	} else {
-		return 0;
-	}
-}
